handle not found error for get_topic crud function and create is_topic_creator utility and then set permissions to topic routes

+++ forum/routes.py
@@ -1,18 +1,15 @@
-from fastapi import APIRouter, Body, Depends
+from fastapi import APIRouter, Body, Depends, HTTPException
 from typing import List
 from sqlalchemy.orm import Session
-from jose import jwt
-from . import crud, schemas
-from auth.dependencies import oauth2_scheme
+from . import crud, schemas, utils
+from auth.dependencies import oauth2_scheme, get_current_active_user
 from db_config import get_db
-import settings
 
 routes = APIRouter(prefix="/topics")
 
 @routes.post("", response_model=schemas.TopicResponse, status_code=201)
-def create_topic(topic: schemas.TopicCreate, db: Session = Depends(get_db), token: str = Depends(oauth2_scheme)):
-    user_id = jwt.decode(token, settings.SECRET_KEY, algorithms=["HS256"])["sub"]
-    return crud.create_topic(db=db, topic=topic, user_id=user_id)
+def create_topic(topic: schemas.TopicCreate, db: Session = Depends(get_db), token: str = Depends(oauth2_scheme), current_user: int = Depends(get_current_active_user)):
+    return crud.create_topic(db=db, topic=topic, user_id=current_user.id)
 
 @routes.get("", response_model=List[schemas.TopicListResponse], status_code=200)
 def topics(db: Session = Depends(get_db)):
@@ -23,12 +20,18 @@ def topic(topic_id: int, db: Session = Depends(get_db)):
     return crud.get_topic(db=db, topic_id=topic_id)
 
 @routes.delete("/{topic_id}", status_code=204)
-def delete_topic(topic_id: int, db: Session = Depends(get_db), token: str = Depends(oauth2_scheme)):
-    return crud.delete_topic(db=db, topic_id=topic_id)
+def delete_topic(topic_id: int, db: Session = Depends(get_db), token: str = Depends(oauth2_scheme), current_user: int = Depends(get_current_active_user)):
+    if utils.is_topic_creator(topic_id=topic_id, user_id=current_user.id, db=db):
+        return crud.delete_topic(db=db, topic_id=topic_id)
+    else:
+        raise HTTPException(status_code=400, detail="You are not the creator of this topic")
 
 @routes.put("/{topic_id}", status_code=200)
-def update_topic(topic_id: int, db: Session = Depends(get_db), topic: schemas.TopicUpdate = Body(), token: str = Depends(oauth2_scheme)):
-    return crud.update_topic(db=db, topic_id=topic_id, topic_schema=topic)
+def update_topic(topic_id: int, db: Session = Depends(get_db), topic: schemas.TopicUpdate = Body(), token: str = Depends(oauth2_scheme), current_user: int = Depends(get_current_active_user)):
+    if utils.is_topic_creator(topic_id=topic_id, user_id=current_user.id, db=db):
+        return crud.update_topic(db=db, topic_id=topic_id, topic_schema=topic)
+    else:
+        raise HTTPException(status_code=400, detail="You are not the creator of this topic")
 
 @routes.patch("/{topic_id}/voteup", status_code=200)
 def voteup_topic(topic_id: int, db: Session = Depends(get_db), token: str = Depends(oauth2_scheme)):
@@ -39,7 +42,7 @@ def votedown_topic(topic_id: int, db: Session = Depends(get_db), token: str = De
     return crud.votedown_topic(db=db, topic_id=topic_id)
 
 @routes.post("/{topic_id}/answers", response_model=schemas.AnswerResponse, status_code=201)
-def create_answer(topic_id: int, answer: schemas.AnswerCreate, db: Session = Depends(get_db), token: str = Depends(oauth2_scheme)):
+def create_answer(topic_id: int, answer: schemas.AnswerCreate, db: Session = Depends(get_db), token: str = Depends(oauth2_scheme), current_user: int = Depends(get_current_active_user)):
     return crud.create_answer(db=db, answer=answer, topic_id=topic_id)
 
 @routes.get("/{topic_id}/answers", response_model=List[schemas.AnswerResponse], status_code=200)
